Project_Vehicle_Tracking/requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
deep-sort-realtime==1.3.2
filelock==3.16.1
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.10.0
humanfriendly==10.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4  # Replace with the correct version from PyPI or Conda
onnx==1.17.0
onnxruntime==1.19.2
opencv-python-headless==4.10.0  # Use this if you don't need GUI features
packaging==24.2
pandas==2.2.3
pillow==11.0.0
protobuf==5.28.3
psutil==6.1.0
py-cpuinfo==9.0.0
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
sympy==1.13.1
torch==2.5.1
torchvision==0.20.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
ultralytics==8.3.28
ultralytics-thop==2.0.11
urllib3==2.2.3
zipp==3.20.2
